refactor-ensure-apprise-passes
