+++ packages/eightballer/agents/derive_arbitrage_agent/aea-config.yaml
@@ -8,7 +8,7 @@ fingerprint:
   tests/test_derive_trader_agent.py: bafybeif4ycyf6kcisfxm6gr224lqxhei4dwk7apfoux4vvpmtoots3ddri
 fingerprint_ignore_patterns: []
 connections:
-- eightballer/apprise:0.1.0:bafybeibirjy3vqzv6grif54ub3tnyn2nreybmcgeorbkiu4ho35jg5cy2u
+- eightballer/apprise_wrapper:0.1.0:bafybeifx733pvk25u62jfanpmcrbw4mktziclvrlwsjtoqhhnk5jmhpzmq
 - eightballer/ccxt_wrapper:0.1.0:bafybeiej6kthjdnrzbz4do5gcbnfdibhxed6ekgx2ywhypo254lzj4nwlu
 - eightballer/dcxt:0.1.0:bafybeicw52cswrbi256okgd43rekm7ens2nuuojcrpelbuappma2v3srpa
 - eightballer/http_client:0.1.0:bafybeiaz5auftwxpt4czrmeeesggqlkc2kosmetq6adrebeu6g7bkhqc2u
@@ -48,7 +48,7 @@ protocols:
 skills:
 - eightballer/abstract_round_abci:0.1.0:bafybeigd3jef57azx7fl3j2dey4eywuqppzpm2ngl3gxci63koqvjsm3re
 - eightballer/prometheus:0.1.0:bafybeicl5i7e467aowfarke4bbyixo2dggar276njmvyuwbsby5pxshhtu
-- eightballer/simple_fsm:0.1.0:bafybeibyb4yljnyfkq7xxeczil4lvubezwzhg3xwqdo6k2i6g4m4mbq2qq
+- eightballer/simple_fsm:0.1.0:bafybeidumg7dffo3ftsstj5acg6cbsenziglyqaznsfzslaxj4tg7crbna
 - eightballer/trading_state:0.1.0:bafybeif7k2vzte7oopana6i2twchxf53qv3knvcsgomdzuubt5ya24fwry
 - valory/abstract_abci:0.1.0:bafybeigiri3x4am542wtw3puasr4rtssvdk3kpu23kmdr6jskommbeoiya
 customs:
@@ -111,7 +111,7 @@ public_id: eightballer/ccxt_wrapper:0.1.0:bafybeidyrx74rtv7x3h757y44cc5xle6crnnk
 type: connection
 config: {}
 ---
-public_id: eightballer/apprise:0.1.0
+public_id: eightballer/apprise_wrapper:0.1.0
 type: connection
 config:
   endpoints:
