+++ packages/eightballer/connections/apprise_wrapper/connection.yaml
@@ -1,4 +1,4 @@
-name: apprise
+name: apprise_wrapper
 author: eightballer
 version: 0.1.0
 type: connection
@@ -8,11 +8,11 @@ description: This is a connection to send notifications using the Apprise librar
 license: Apache-2.0
 aea_version: '>=1.0.0, <2.0.0'
 fingerprint:
-  README.md: bafybeibopk4k4hvinxywoayselnr65an5m7t3jc6hpfa6s3eov42gu4o4m
+  README.md: bafybeig2mnxn6qkckwvfjca4ary3ctgwnrbhh2b7todxgqrkedcpxzcnt4
   __init__.py: bafybeiedjyvl6kbspjhvzfiz6b6uhrep4nyolybcuug3645cbcfgur4fnu
-  connection.py: bafybeib5va5nycw4zex6domdsq46e5uykmqenh6aa257ktkdyji3c5qleu
+  connection.py: bafybeigx7jz4h2trt3baytmdjlezgap3277nonj6gaxdm4vpdxrj5g427u
   tests/__init__.py: bafybeic7sbvhv3hzf4vjgycb6oqogqknle72kag2fanocp2ql7b33qx77a
-  tests/test_connection.py: bafybeifaowiinbavbm3xgksuuwdqh4by6a44ezeoa6zrzq33lhsceq5xai
+  tests/test_connection.py: bafybeifqwp6upbu5d4ckjdoc4dk3fycs4ipjvbujsnxf3lgyasddym224a
 fingerprint_ignore_patterns: []
 connections: []
 protocols:
