+++ packages/eightballer/skills/simple_fsm/strategy.py
@@ -29,10 +29,10 @@ from aea.skills.base import Model
 from aea.configurations.base import PublicId
 
 from packages.eightballer.protocols.orders.custom_types import Order
-from packages.eightballer.connections.apprise.connection import CONNECTION_ID as APPRISE_PUBLIC_ID
 from packages.eightballer.skills.abstract_round_abci.models import FrozenMixin
 from packages.eightballer.protocols.user_interaction.message import UserInteractionMessage
 from packages.eightballer.protocols.user_interaction.dialogues import UserInteractionDialogues
+from packages.eightballer.connections.apprise_wrapper.connection import CONNECTION_ID as APPRISE_PUBLIC_ID
 
 
 TZ = datetime.datetime.now().astimezone().tzinfo
