+++ packages/packages.json
@@ -25,18 +25,18 @@
         "contract/eightballer/amb_gnosis_eth_xdai/0.1.0": "bafybeiamyb7fq3as3wcv3dvlwm7ilt5clzfsyrlv7uihn2yipvv6h2cn5a",
         "connection/eightballer/dcxt/0.1.0": "bafybeicw52cswrbi256okgd43rekm7ens2nuuojcrpelbuappma2v3srpa",
         "connection/eightballer/ccxt_wrapper/0.1.0": "bafybeiej6kthjdnrzbz4do5gcbnfdibhxed6ekgx2ywhypo254lzj4nwlu",
-        "connection/eightballer/apprise/0.1.0": "bafybeibirjy3vqzv6grif54ub3tnyn2nreybmcgeorbkiu4ho35jg5cy2u",
+        "connection/eightballer/apprise_wrapper/0.1.0": "bafybeifx733pvk25u62jfanpmcrbw4mktziclvrlwsjtoqhhnk5jmhpzmq",
         "skill/eightballer/dex_data_retrieval/0.1.0": "bafybeibj5tu3rf7gpf3evopnu7xpqqknpuzbkt6cu6pzzc43tmkzylfdiu",
         "skill/eightballer/reporting/0.1.0": "bafybeiaqoa7gkubrwpjrelvnixiwusuonjvuvqnran6zixrhgr34l2az54",
         "skill/eightballer/chained_dex_app/0.1.0": "bafybeifdpgoxnunruv4572am5ufwg4fk3stnioio3bbabzlsixeuo5govy",
-        "skill/eightballer/simple_fsm/0.1.0": "bafybeibyb4yljnyfkq7xxeczil4lvubezwzhg3xwqdo6k2i6g4m4mbq2qq",
+        "skill/eightballer/simple_fsm/0.1.0": "bafybeidumg7dffo3ftsstj5acg6cbsenziglyqaznsfzslaxj4tg7crbna",
         "skill/eightballer/gnosis_bridging_abci_app/0.1.0": "bafybeihwbxlza3j3v2lruhuvbeqljf45ruwnntmrupxdbtj7kdcucf7kby",
         "skill/eightballer/trading_state/0.1.0": "bafybeif7k2vzte7oopana6i2twchxf53qv3knvcsgomdzuubt5ya24fwry",
         "agent/eightballer/chained_dex_app/0.1.0": "bafybeieossuhuk2hdul63wkjptwwkfne6izbqegqdoog32bftl6a72hw2m",
-        "agent/eightballer/trader/0.1.0": "bafybeigjzoizlze2kueq4muqpmfzkvdtrmiayim374vrg5j5rwti7ioagy",
+        "agent/eightballer/trader/0.1.0": "bafybeihozffm2ss2c663uzzklzisnvzvdlpj7qqlqxqctqybbjpmudcfim",
         "agent/eightballer/bridging_agent/0.1.0": "bafybeigkngh76memzf7bumrumvhsg3m4xuf3bjel4l3bx5wd3yvcdhk4va",
-        "agent/eightballer/derive_arbitrage_agent/0.1.0": "bafybeigbtvsqjrphgmcvsjtd5aghkakb6alwwninmddijfrmuacsw4swqm",
-        "service/eightballer/cex_dex_arbitrage/0.1.0": "bafybeia3zqxki53yuvybr7km7vkijpfud57w7h5w4rmxtjrxqffk3ezruq"
+        "agent/eightballer/derive_arbitrage_agent/0.1.0": "bafybeigsymnl3bbc32jmgt6jovvhmkxyhwrqjr6zdzeylw5oycowy2boy4",
+        "service/eightballer/cex_dex_arbitrage/0.1.0": "bafybeihkxryxsczdnbho7jwfosd4dqtgyslqd3vum4eiao56zgybs6jzpq"
     },
     "third_party": {
         "protocol/valory/abci/0.1.0": "bafybeiatodhboj6a3p35x4f4b342lzk6ckxpud23awnqbxwjeon3k5y36u",
